requirements.txt
alembic==1.11.1
beautifulsoup4==4.12.2
Flask==2.3.2
Flask_Admin==1.6.1
flask_babel==4.0.0
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.1.1
Flask_WTF==1.1.1
MySQL-python==1.2.5
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.18
Werkzeug==2.3.6
WTForms==3.0.1
